Results from my benchmark:

       user     system      total        real
with CollectionRender rendered 1 time:  0.007166   0.001877   0.009043 (  0.011686)
without CollectionRender rendered 1 times:  0.000740   0.000192   0.000932 (  0.001078)
with CollectionRender rendered 100 times:  0.021146   0.001034   0.022180 (  0.022362)
without CollectionRender rendered 100 times:  0.028972   0.001244   0.030216 (  0.030332)
with CollectionRender order 50000 collection:  0.008611   0.000440   0.009051 (  0.009060)
without CollectionRender order 50000 collection:  0.842578   0.018036   0.860614 (  0.863523)